refactor file manager
