+++ src/functions/fs/deleteFile.js
