+++ src/functions/fs/readfile.js
