+++ src/functions/ls.js
@@ -3,12 +3,8 @@ import fs from 'fs/promises';
 const printList = async (directory) => {
   try {
     const entries = await fs.readdir(directory, { withFileTypes: true });
-    const folders = entries.filter((file) => file.isDirectory());
-    const files = entries.filter((file) => file.isFile());
-
-    const list = folders.concat(files);
     console.table(
-      list.map((item) => ({
+      entries.map((item) => ({
         Name: item.name,
         Type: item.isDirectory() ? 'directory' : 'file',
       }))
